feat: use nostale bindings instead of nossmooth core

+++ Core/NosSmooth.Core/Commands/WalkCommand.cs
@@ -13,4 +13,4 @@ namespace NosSmooth.Core.Commands;
 /// <param name="TargetX">The x coordinate of the target position to move to.</param>
 /// <param name="TargetY">The y coordinate of the target position to move to.</param>
 /// <param name="CancelOnUserMove">Whether to cancel the walk when the user clicks to move somewhere.</param>
-public record WalkCommand(int TargetX, int TargetY, bool CancelOnUserMove = true) : ICommand;
+public record WalkCommand(ushort TargetX, ushort TargetY, bool CancelOnUserMove = true) : ICommand;
